Refactor data modules
