+++ main-packages/common-syntax/lib/Minicute/Data/Common/ExpressionLevel.hs
@@ -1,0 +1,22 @@
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DeriveLift #-}
+module Minicute.Data.Common.ExpressionLevel
+  ( ExpressionLevel( .. )
+  ) where
+
+import Data.Data ( Data, Typeable )
+import GHC.Generics ( Generic )
+import Language.Haskell.TH.Syntax ( Lift )
+
+data ExpressionLevel
+  = MC -- ^ miniCUTE
+  | LLMC -- ^ Lambda lifted miniCUTE
+  deriving ( Generic
+           , Typeable
+           , Data
+           , Lift
+           , Eq
+           , Ord
+           , Show
+           )
