+++ main-packages/llvm-generator/test/Minicute/Transpilers/LLVMSpec.hs
@@ -14,7 +14,7 @@ import Data.Tuple.Extra
 import LLVM.IRBuilder
 import Minicute.Transpilers.LLVM
 import Minicute.Transpilers.LLVM.Constants
-import Minicute.Utils.TH.GMachine
+import Minicute.Utils.GMachine.TH
 
 import qualified LLVM.AST as AST
 import qualified LLVM.AST.Type as ASTT
