+++ main-packages/minicute-optimizer/test/Minicute/Transpilers/Optimizers/SimpleArithmeticSpec.hs
@@ -10,7 +10,7 @@ import Control.Monad
 import Data.Tuple.Extra
 import Minicute.Data.Minicute.Program
 import Minicute.Transpilers.Optimizers.SimpleArithmetic
-import Minicute.Utils.TH.Minicute
+import Minicute.Utils.Minicute.TH
 
 spec :: Spec
 spec = do
